docs: add Parser and Lexer grammars
Add SlowLangV2Parser and SlowLangV2Lexer grammars to /grammars folder.
The parser grammar defines the production rules for the SlowLang programming language.
The lexer grammar defines the initial set of tokens for the SlowLang programming language.
The grammars are in .g4 format which is used by antlr4 to generate source code for the parser and lexer.