feat: add Compiler
Introduce Compiler class which uses a file reader, lexer and parser to read, tokenize and parse SlowLang source code.
The Compiler outputs a ParseTree (antlr4 class which represents a CST - Concrete Syntax Tree, a.k.a Parse Tree).
Add FileReader class which reads a .txt file from the main directory of the project and outputs a String.