feat: add ExpressionTranslator feature
Add ExpressionTranslator class which translates a list of parsed expressions in post order notation into bytecode.
In the current version of the class 'bytecode' refers to an array of integers. In future versions this class will output proper bytecode.

Add TreeWalker which converts an antlr4 ParseTree object to a list of tokens in post order notation.
Add InfixToPostfixConverter which converts a list of tokens in standard infix notation to postfix notation.